// src/lib/api/generated/model/appointmentStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AppointmentStatus = {
  requested: 'requested',
  confirmed: 'confirmed',
  in_progress: 'in_progress',
  done: 'done',
  no_show: 'no_show',
  cancelled: 'cancelled',
} as const;